backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic
python-dotenv
alembic
psycopg2-binary
passlib[bcrypt]
python-jose[cryptography]
